Jinha 휴가 전자결재로직

+++ HiveWorks/src/main/java/com/dna/hiveworks/model/dto/edoc/ElectronicDocument.java
@@ -42,7 +42,7 @@ public class ElectronicDocument implements Serializable{
 	private String edocContent;
 	private Date edocStartDate;
 	private Date edocEndDate;
-	private int edocVacationCount;
+	private double edocVacationCount;
 	private String edocStatus;
 	private int creater;
 	private Date createDate;
